+++ include/daw/daw_logic.h
@@ -13,13 +13,15 @@
 namespace daw {
 	/// Explicitly do a non-short circuiting logical and with bitwise and
 	template<typename Bool, typename... Bools>
-	DAW_ATTRIB_INLINE constexpr bool nsc_and( Bool lhs, Bools... rhs ) noexcept {
+	[[nodiscard]] DAW_ATTRIB_INLINE constexpr bool
+	nsc_and( Bool lhs, Bools... rhs ) noexcept {
 		return ( static_cast<bool>( rhs ) & ... & static_cast<bool>( lhs ) );
 	}
 
 	/// Explicitly do a non-short circuiting logical or with bitwise or
 	template<typename Bool, typename... Bools>
-	DAW_ATTRIB_INLINE constexpr bool nsc_or( Bool lhs, Bools... rhs ) noexcept {
+	[[nodiscard]] DAW_ATTRIB_INLINE constexpr bool
+	nsc_or( Bool lhs, Bools... rhs ) noexcept {
 		return ( static_cast<bool>( rhs ) | ... | static_cast<bool>( lhs ) );
 	}
 } // namespace daw
