+++ include/daw/daw_swap.h
@@ -12,7 +12,6 @@
 #include "cpp_17.h"
 #include "daw_exchange.h"
 #include "daw_move.h"
-#include "daw_swap.h"
 
 #include <array>
 #include <optional>
