+++ tests/daw_iterator_indexed_iterator_test.cpp
@@ -7,6 +7,7 @@
 //
 
 #include "daw/cpp_17.h"
+#include "daw/cpp_17_iterator.h"
 #include "daw/iterator/daw_indexed_iterator.h"
 
 #include <array>
