+++ tests/daw_visit_test.cpp
@@ -109,7 +109,8 @@ struct T12 {};
 struct T13 {};
 struct T14 {};
 struct T15 {};
-struct T16 {};
+struct T16 {
+};
 constexpr std::variant<T0, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12,
                        T13, T14, T15, T16>
 foobar( int n ) {
@@ -153,9 +154,6 @@ foobar( int n ) {
 int foofoo( std::variant<T0, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12,
                          T13, T14, T15, T16>
               v ) {
-	if( v.valueless_by_exception( ) ) {
-		return -1;
-	}
 	return daw::visit_nt(
 	  v,
 	  []( T0 ) {
